% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-members.r
\name{as_tibble.ppmembers}
\alias{as_tibble.ppmembers}
\alias{as_tibble.ppmember}
\alias{as_tibble.ppmembers_new}
\alias{as_tibble.ppmembers_location}
\alias{as_tibble.ppmembers_leaving}
\alias{as_tibble.ppmember_votes}
\alias{as_tibble.ppmember_bills}
\title{Convert json object to a dataframe}
\usage{
\method{as_tibble}{ppmembers}(x, ...)

\method{as_tibble}{ppmember}(x, ...)

\method{as_tibble}{ppmembers_new}(x, ...)

\method{as_tibble}{ppmembers_location}(x, ...)

\method{as_tibble}{ppmembers_leaving}(x, ...)

\method{as_tibble}{ppmember_votes}(x, ...)

\method{as_tibble}{ppmember_bills}(x, ...)
}
\arguments{
\item{x}{list object}

\item{...}{other parameters passed to \code{\link[tibble]{as_tibble}}:}
}
\description{
Convert json object to a dataframe
}
