% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-bills.r
\name{bill_cosponsors}
\alias{bill_cosponsors}
\title{Get Cosponsors for a Specific Bill}
\usage{
bill_cosponsors(bill_id, congress = 116L, page = 1L)
}
\arguments{
\item{bill_id}{a bill slug, for example hr4881 - these can be found in the recent bill response.}

\item{congress}{105-116}

\item{page}{page of results}
}
\description{
Get information about the cosponsors of a particular bill
}
\examples{
\dontrun{
bill_cosponsors("hr4249", 114)
}

}
\references{
\url{https://projects.propublica.org/api-docs/congress-api/bills/#get-cosponsors-for-a-specific-bill}
}
