% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-bills.r
\name{bill_related}
\alias{bill_related}
\title{Get Related Bills for a Specific Bill}
\usage{
bill_related(bill_id, congress = 116L, page = 1L)
}
\arguments{
\item{bill_id}{a bill slug, for example hr4881 - these can be found in the recent bill response.}

\item{congress}{105-116}

\item{page}{page of results}
}
\description{
Use this request type to get Library of Congress-identified related bills for
a particular bill. This request returns the 20 most recent results and
supports paginated requests.
}
\examples{
\dontrun{
bill_related("hr3219", 115)
}

}
\references{
\url{https://projects.propublica.org/api-docs/congress-api/bills/#get-related-bills-for-a-specific-bill}
}
