% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-bills.r
\name{bill_subjects}
\alias{bill_subjects}
\title{Get Subjects for a Specific Bill}
\usage{
bill_subjects(bill_id, congress = 116L, page = 1L)
}
\arguments{
\item{bill_id}{a bill slug, for example hr4881 - these can be found in the recent bill response.}

\item{congress}{105-116}

\item{page}{page of results}
}
\description{
Use this request type to get Library of Congress-assigned subjects about a
particular bill. This request returns the 20 most recent results and supports
paginated requests.
}
\examples{
\dontrun{
bill_subjects("hr2810", 115)
}

}
\references{
\url{https://projects.propublica.org/api-docs/congress-api/bills/#get-subjects-for-a-specific-bill}
}
