% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-bills.r
\name{bills_recent}
\alias{bills_recent}
\title{Get Recent Bills}
\usage{
bills_recent(chamber = c("house", "senate"), congress = 116,
  type = c("introduced", "updated", "active", "passed", "enacted",
  "vetoed"), page = 1L)
}
\arguments{
\item{chamber}{house or senate}

\item{congress}{105-116}

\item{type}{introduced, updated, active, passed, enacted, or vetoed}

\item{page}{page of results}
}
\description{
Use this request type to get summaries of the 20 most recent bills by type.
For the current Congress, “recent bills” can be one of four types
(see the descriptions below). For previous Congresses, “recent bills” means
the last 20 bills of that Congress. In the responses, an \code{active} value of true
means that the bill has seen action beyond introduction and committee
referral. Requests include a \code{chamber} value; to get recent bills from the
House and Senate, use both as the value. You can paginate through bills using
the \code{offset} querystring parameter that accepts multiples of 20. Bills before
the 113th Congress (prior to 2013) have fewer attribute values than those
from the 113th Congress onward, because the more recent bill data comes from
the bulk data provided by the Government Publishing Office. Details for the
older bills came from scraping Thomas.gov, the former congressional site of
the Library of Congress.
}
\examples{
\dontrun{
bills_recent("house", 116, "enacted")
}

}
\references{
\url{https://projects.propublica.org/api-docs/congress-api/bills/#search-bills}
}
