% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-bills.r
\name{bills_recent_subject}
\alias{bills_recent_subject}
\title{Recent Bills by a Specific Subject}
\usage{
bills_recent_subject(subject, page = 1L)
}
\arguments{
\item{subject}{A slug version of a legislative subject}

\item{page}{page of results}
}
\description{
Use this request type to get the 20 most recently updated bills for a specific
legislative subject. Results can include more than one Congress.
}
\examples{
\dontrun{
bills_recent_subject("meat")
}

}
\references{
\url{https://projects.propublica.org/api-docs/congress-api/bills/#get-recent-bills-by-a-specific-subject}
}
\keyword{internal}
