% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-bills.r
\name{bills_search_subjects}
\alias{bills_search_subjects}
\title{Search a Specific Bill Subject}
\usage{
bills_search_subjects(query, page = 1L)
}
\arguments{
\item{query}{a word or phrase to search}

\item{page}{page of results}
}
\description{
Use this request type to search for bill subjects that contain a specified
term.
}
\examples{
\dontrun{
bills_search_subjects("climate")
}

}
\references{
\url{https://projects.propublica.org/api-docs/congress-api/bills/#get-a-specific-bill-subject}
}
