% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-bills.r
\name{bills_upcoming}
\alias{bills_upcoming}
\title{Get Upcoming Bills}
\usage{
bills_upcoming(chamber = c("house", "senate"))
}
\arguments{
\item{chamber}{house or senate}
}
\description{
Use this request type to get details on bills that may be considered by the
House or Senate in the near future, based on scheduled published or announced
by congressional leadership. The bills and their potential consideration are
taken from the \href{http://docs.house.gov/floor/}{House Majority Leader} and
floor updates from
\href{http://www.republicans.senate.gov/public/index.cfm/floor-updates}{Senate Republicans}.
The responses include a \code{legislative_day} attribute which is the earliest the
bills could be considered, and a \code{range} attribute that indicates whether the
bill information comes from a weekly schedule or a daily one. Combine the two
for the best sense of when a bill might come up for consideration. For Senate
bills, the response includes a \code{context} attribute reproducing the sentence that
includes mention of the bill. These responses omit bills that have not yet been
assigned a bill number or introduced, and additional bills may be considered
at any time.
}
\examples{
\dontrun{
bills_upcoming("house")
}

}
\references{
\url{https://projects.propublica.org/api-docs/congress-api/bills/#get-upcoming-bills}
}
