% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.r
\name{congress_api}
\alias{congress_api}
\title{Congress API}
\usage{
congress_api(path, query = NULL)
}
\arguments{
\item{path}{path to endpoint}
}
\description{
Make request to ProPublic Congress API.
}
\keyword{internal}
