% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{extract_single_result}
\alias{extract_single_result}
\title{Extract a single result from a json list}
\usage{
extract_single_result(x)
}
\arguments{
\item{x}{a list}
}
\description{
This is useful when we know that the resulting json
will contain exactly one result
}
