% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-members.r
\name{member}
\alias{member}
\title{Get a Specific Member}
\usage{
member(member_id)
}
\arguments{
\item{member_id}{The ID of the member to retrieve; it is assigned by the \href{http://bioguide.congress.gov/biosearch/biosearch.asp}{Biographical Directory of the United States Congress} or can be retrieved from a member list request.}
}
\description{
Get biographical and Congressional role information for a particular member
of Congress
}
\examples{
\dontrun{
member("O000172")
}

}
\references{
\url{https://projects.propublica.org/api-docs/congress-api/members/#get-a-specific-member}
}
