% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-members.r
\name{member_bills}
\alias{member_bills}
\title{Get Bills Cosponsored by a Specific Member}
\usage{
member_bills(member_id, type = c("cosponsored", "withdrawn"),
  page = 1L)
}
\arguments{
\item{member_id}{The ID of the member to retrieve; it is assigned by the \href{http://bioguide.congress.gov/biosearch/biosearch.asp}{Biographical Directory of the United States Congress} or can be retrieved from a member list request.}

\item{type}{cosponsored or withdrawn}

\item{page}{page of results}
}
\description{
Get the 20 most recent bill cosponsorships for a particular member, either
bills cosponsored or bills where cosponsorship was withdrawn
}
\examples{
\dontrun{
member_bills("O000172")
}
}
