% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-members.r
\name{member_votes}
\alias{member_votes}
\title{Get a Specific Member’s Vote Positions}
\usage{
member_votes(member_id, page = 1L)
}
\arguments{
\item{member_id}{The ID of the member to retrieve; it is assigned by the \href{http://bioguide.congress.gov/biosearch/biosearch.asp}{Biographical Directory of the United States Congress} or can be retrieved from a member list request.}

\item{page}{page of results}
}
\description{
Get the most recent vote positions for a specific member of the House of Representatives or Senate
}
\examples{
\dontrun{
member_votes("O000172")
}

}
\references{
\url{https://projects.propublica.org/api-docs/congress-api/members/#get-a-specific-members-vote-positions}
}
