% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-members.r
\name{members}
\alias{members}
\title{Lists of Members}
\usage{
members(chamber = c("house", "senate"), congress = 116)
}
\arguments{
\item{chamber}{house or senate}

\item{congress}{102-116 for House, 80-116 for Senate}
}
\description{
Get a list of members of a particular chamber in a particular Congress.
The results include all members who have served in that congress and chamber,
 including members who are no longer in office. To filter the list to only
 active members (or to see members who have left), use the in_office boolean
 attribute.
}
\examples{
\dontrun{
members("house", 116)
}
}
