% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-members.r
\name{members_leaving}
\alias{members_leaving}
\title{Get Members Leaving Office}
\usage{
members_leaving(chamber = c("house", "senate"), congress = 116,
  page = 1L)
}
\arguments{
\item{chamber}{house or senate}

\item{congress}{102-116 for House, 80-116 for Senate}

\item{page}{page of results}
}
\description{
get a list of members who have left the Senate or House or have announced plans to do so.
}
\examples{
\dontrun{
members_leaving("senate", 116)
}

}
\references{
\url{https://projects.propublica.org/api-docs/congress-api/members/#get-members-leaving-office}
}
