% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-members.r
\name{members_location}
\alias{members_location}
\title{Get Current Members by State/District}
\usage{
members_location(chamber = c("house", "senate"), state = NULL,
  district = NULL)
}
\arguments{
\item{chamber}{house or senate}

\item{state}{two-letter state abbreviation (ex. "TX")}

\item{district}{district number (ex. 1)}
}
\description{
Get biographical and Congressional role information for a particular member of Congress
}
\examples{
\dontrun{
members_location("senate", "TX")
}
}
