% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-members.r
\name{members_new}
\alias{members_new}
\title{Get new members}
\usage{
members_new()
}
\description{
Get a list of the 20 most recent new members of the current Congress.
}
\examples{
\dontrun{
members_new()
}
}
