% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.r
\name{new_congress}
\alias{new_congress}
\title{Create a new congress object}
\usage{
new_congress(json, path)
}
\arguments{
\item{json}{list data from json api}

\item{path}{path to location of data}
}
\description{
Create a new congress object
}
\keyword{internal}
