% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.r
\name{set_ppclass}
\alias{set_ppclass}
\title{Set class of pp object}
\usage{
set_ppclass(x, name)
}
\arguments{
\item{x}{a ppcongres_api object}

\item{name}{the name of the new class}
}
\description{
Set class of pp object
}
