% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-bills.r
\name{as_tibble.ppbills_search}
\alias{as_tibble.ppbills_search}
\alias{as_tibble.ppbills_recent}
\alias{as_tibble.ppbills_search_subjects}
\alias{as_tibble.ppbills_upcoming}
\alias{as_tibble.ppbill}
\alias{as_tibble.ppbill_amendments}
\alias{as_tibble.ppbill_subjects}
\title{Convert json object to a dataframe}
\usage{
\method{as_tibble}{ppbills_search}(x, ...)

\method{as_tibble}{ppbills_recent}(x, ...)

\method{as_tibble}{ppbills_search_subjects}(x, ...)

\method{as_tibble}{ppbills_upcoming}(x, ...)

\method{as_tibble}{ppbill}(x, ...)

\method{as_tibble}{ppbill_amendments}(x, ...)

\method{as_tibble}{ppbill_subjects}(x, ...)
}
\arguments{
\item{x}{list object}

\item{...}{other parameters passed to \code{\link[tibble]{as_tibble}}:}
}
\description{
Convert json object to a dataframe
}
